+++ Examples/Python/src/ExaTrkXTrackFinding.cpp
@@ -58,7 +58,7 @@ using namespace py::literals;
 namespace ActsPython {
 
 void addExaTrkXTrackFinding(Context &ctx) {
-  auto [m, mex] = ctx.get("main", "examples");
+  auto &mex = ctx.get("examples");
 
   {
     using C = Acts::GraphConstructionBase;
