+++ Python/Core/src/GeometryGen3.cpp
@@ -6,9 +6,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "Acts/Geometry/Blueprint.hpp"
-
 #include "Acts/Definitions/Units.hpp"
+#include "Acts/Geometry/Blueprint.hpp"
 #include "Acts/Geometry/BlueprintNode.hpp"
 #include "Acts/Geometry/ContainerBlueprintNode.hpp"
 #include "Acts/Geometry/CylinderVolumeStack.hpp"
@@ -217,7 +216,7 @@ void pseudoNavigation(const Acts::TrackingGeometry& trackingGeometry,
 
 }  // namespace
 
-void addBlueprint(Context& ctx) {
+void addGeometryGen3(Context& ctx) {
   using Acts::Experimental::Blueprint;
   using Acts::Experimental::BlueprintNode;
   using Acts::Experimental::BlueprintOptions;
